+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* src() should not return blob URLs with MSE source handlers ([view](https://github.com/videojs/video.js/pull/2271))
 
 --------------------